+++ buildSrc/src/main/kotlin/Common.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Rick Busarow
+ * Copyright (C) 2021 Rick Busarow
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,13 +13,19 @@
  * limitations under the License.
  */
 
-@file:Suppress("LongMethod", "TopLevelPropertyNaming")
-
 import org.gradle.api.*
+import org.gradle.api.tasks.testing.*
 import org.gradle.kotlin.dsl.*
 import org.jetbrains.kotlin.gradle.tasks.*
 
 fun Project.common() {
+
+  tasks.withType<Test> {
+
+    useJUnitPlatform {
+      includeEngines("junit-vintage", "junit-jupiter")
+    }
+  }
 
   tasks.withType<KotlinCompile>()
     .configureEach {
@@ -32,27 +38,19 @@ fun Project.common() {
       }
     }
 
-// force update all transitive dependencies (prevents some library leaking an old version)
+  // force update all transitive dependencies (prevents some library leaking an old version)
   configurations.all {
     resolutionStrategy {
       force(
-        Libs.Kotlin.reflect,
-        // androidx is currently leaking coroutines 1.1.1 everywhere
-        Libs.Kotlinx.Coroutines.core,
-        Libs.Kotlinx.Coroutines.test,
-        Libs.Kotlinx.Coroutines.android,
-        // prevent dependency libraries from leaking their own old version of this library
-        Libs.RickBusarow.Dispatch.core,
-        Libs.RickBusarow.Dispatch.detekt,
-        Libs.RickBusarow.Dispatch.espresso,
-        Libs.RickBusarow.Dispatch.lifecycle,
-        Libs.RickBusarow.Dispatch.lifecycleExtensions,
-        Libs.RickBusarow.Dispatch.viewModel,
-        Libs.RickBusarow.Dispatch.Test.core,
-        Libs.RickBusarow.Dispatch.Test.jUnit4,
-        Libs.RickBusarow.Dispatch.Test.jUnit5
+        "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.3",
+        "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.4.3",
+        "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.4.3"
       )
+      eachDependency {
+        when {
+          requested.group == "org.jetbrains.kotlin" -> useVersion("1.5.0")
+        }
+      }
     }
   }
-
 }