+++ dispatch-android-espresso/src/test/java/dispatch/android/espresso/IdlingDispatcherProviderRuleTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Rick Busarow
+ * Copyright (C) 2021 Rick Busarow
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -47,9 +47,7 @@ class IdlingDispatcherProviderRuleTest {
     verify { registry.register(idlingDispatcherProvider.main.counter) }
     verify { registry.register(idlingDispatcherProvider.mainImmediate.counter) }
     verify { registry.register(idlingDispatcherProvider.unconfined.counter) }
-
   }
-
 }
 
 class EspressoMockingWrapper : TestWatcher() {