+++ dispatch-android-lifecycle-extensions/src/main/java/dispatch/android/lifecycle/internal/DispatchLifecycleScopeStore.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Rick Busarow
+ * Copyright (C) 2021 Rick Busarow
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -55,7 +55,7 @@ internal object DispatchLifecycleScopeStore : LifecycleEventObserver {
          */
         (map as ConcurrentMap).atomicGetOrPut(lifecycle) { bindLifecycle(lifecycle) }
       }
-      else                        -> {
+      else -> {
         synchronized(map) {
           map.getOrPut(lifecycle) { bindLifecycle(lifecycle) }
         }
@@ -75,5 +75,4 @@ internal object DispatchLifecycleScopeStore : LifecycleEventObserver {
 
     return scope
   }
-
 }