+++ dispatch-android-lifecycle/src/main/java/dispatch/android/lifecycle/internal/viewLifecycleCoroutineScope.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Rick Busarow
+ * Copyright (C) 2021 Rick Busarow
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -48,17 +48,17 @@ internal fun bindViewLifecycleCoroutineScope(
         )
 
         viewScope.block()
-
       }
     }.flowOnMainImmediate()
     .launchIn(receiverScope)
 
-  fragment.lifecycle.addObserver(LifecycleEventObserver { _, event ->
-    if (event == Lifecycle.Event.ON_DESTROY) {
-      observerJob.cancel()
+  fragment.lifecycle.addObserver(
+    LifecycleEventObserver { _, event ->
+      if (event == Lifecycle.Event.ON_DESTROY) {
+        observerJob.cancel()
+      }
     }
-  })
+  )
 
   return observerJob
 }
-