+++ dispatch-android-viewmodel/samples/src/test/java/samples/ViewModelScopeFactorySample.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Rick Busarow
+ * Copyright (C) 2021 Rick Busarow
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -78,5 +78,4 @@ class ViewModelScopeFactorySample {
 
   private annotation class Before
   private annotation class After
-
 }