+++ dispatch-core/src/main/java/dispatch/core/Async.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Rick Busarow
+ * Copyright (C) 2021 Rick Busarow
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -104,4 +104,3 @@ public fun <T> CoroutineScope.asyncUnconfined(
   start: CoroutineStart = CoroutineStart.DEFAULT,
   block: suspend CoroutineScope.() -> T
 ): Deferred<T> = async(context + coroutineContext.dispatcherProvider.unconfined, start, block)
-