+++ dispatch-core/src/main/java/dispatch/core/Suspend.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Rick Busarow
+ * Copyright (C) 2021 Rick Busarow
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -95,7 +95,8 @@ public suspend fun <T> withMain(
  * @see withContext
  */
 public suspend fun <T> withMainImmediate(
-  context: CoroutineContext = EmptyCoroutineContext, block: suspend CoroutineScope.() -> T
+  context: CoroutineContext = EmptyCoroutineContext,
+  block: suspend CoroutineScope.() -> T
 ): T {
   val newContext = context + coroutineContext.dispatcherProvider.mainImmediate
   return withContext(newContext, block)