+++ dispatch-detekt/src/main/java/dispatch/detekt/rules/AndroidXLifecycleScope.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Rick Busarow
+ * Copyright (C) 2021 Rick Busarow
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -26,12 +26,12 @@ import org.jetbrains.kotlin.psi.psiUtil.*
  * The AndroidX library uses a hard-coded `Dispatchers.Main` and does not contain a `DispatcherProvider`,
  * and also leaks its pausing behavior.
  */
-class AndroidXLifecycleScope(config: Config = Config.empty) : Rule(config) {
+public class AndroidXLifecycleScope(config: Config = Config.empty) : Rule(config) {
 
   /**
    * @suppress
    */
-  override val issue = Issue(
+  override val issue: Issue = Issue(
     id = "AndroidXLifecycleScope",
     severity = Severity.Defect,
     description = "The AndroidX lifecycleScope leaks its pausing behavior and uses hard-coded dispatchers.",
@@ -41,10 +41,10 @@ class AndroidXLifecycleScope(config: Config = Config.empty) : Rule(config) {
   /**
    * @suppress
    */
-  companion object {
-    val starImportRegex = "androidx\\.lifecycle\\.\\*".toRegex()
-    val explicitImportRegex = "androidx\\.lifecycle\\.lifecycleScope".toRegex()
-    const val functionName = "lifecycleScope"
+  private companion object {
+    private val starImportRegex = "androidx\\.lifecycle\\.\\*".toRegex()
+    private val explicitImportRegex = "androidx\\.lifecycle\\.lifecycleScope".toRegex()
+    private const val functionName = "lifecycleScope"
   }
 
   private val calls = mutableListOf<KtReferenceExpression>()
@@ -71,7 +71,6 @@ class AndroidXLifecycleScope(config: Config = Config.empty) : Rule(config) {
         report(finding)
       }
     }
-
   }
 
   /**
@@ -102,6 +101,3 @@ class AndroidXLifecycleScope(config: Config = Config.empty) : Rule(config) {
     super.visitImportDirective(importDirective)
   }
 }
-
-
-