+++ dispatch-internal-test-android/src/main/java/dispatch/internal/test/android/FakeLifecycle.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Rick Busarow
+ * Copyright (C) 2021 Rick Busarow
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -17,12 +17,16 @@ package dispatch.internal.test.android
 
 import androidx.lifecycle.*
 import kotlinx.coroutines.channels.*
+import kotlinx.coroutines.flow.*
 
 class FakeLifecycle(lifecycleOwner: FakeLifecycleOwner) : Lifecycle() {
 
   val delegate = LifecycleRegistry(lifecycleOwner)
 
-  val observerEvents = Channel<ObserverEvent>(1200)
+  val observerEvents = MutableSharedFlow<ObserverEvent>(
+    replay = 1,
+    onBufferOverflow = BufferOverflow.DROP_OLDEST
+  )
 
   val observerCount: Int get() = delegate.observerCount
 
@@ -35,12 +39,12 @@ class FakeLifecycle(lifecycleOwner: FakeLifecycleOwner) : Lifecycle() {
 
   override fun addObserver(observer: LifecycleObserver) {
     delegate.addObserver(observer)
-    observerEvents.sendBlocking(ObserverEvent.Add(observer))
+    observerEvents.tryEmit(ObserverEvent.Add(observer))
   }
 
   override fun removeObserver(observer: LifecycleObserver) {
     delegate.removeObserver(observer)
-    observerEvents.sendBlocking(ObserverEvent.Remove(observer))
+    observerEvents.tryEmit(ObserverEvent.Remove(observer))
   }
 
   override fun getCurrentState(): State = delegate.currentState
@@ -51,5 +55,4 @@ class FakeLifecycle(lifecycleOwner: FakeLifecycleOwner) : Lifecycle() {
   fun handleLifecycleEvent(event: Event) {
     delegate.handleLifecycleEvent(event)
   }
-
 }