+++ dispatch-sample/src/main/java/dispatch/sample/MainFragment.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 Rick Busarow
+ * Copyright (C) 2021 Rick Busarow
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -57,7 +57,6 @@ class MainFragment : Fragment() {
         }
       }
     }
-
   }
 
   override fun onCreateView(
@@ -91,15 +90,14 @@ class MainFragment : Fragment() {
       // the default dispatcher in this scope is now Dispatchers.IO
       .launchIn(ioDefaultScope)
   }
-
 }
 
 @Suppress("UNCHECKED_CAST")
 inline fun <reified VM : ViewModel> viewModelFactory(crossinline f: () -> VM):
-    ViewModelProvider.Factory =
-  object : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(aClass: Class<T>): T = f() as T
-  }
+  ViewModelProvider.Factory =
+    object : ViewModelProvider.Factory {
+      override fun <T : ViewModel> create(aClass: Class<T>): T = f() as T
+    }
 
 inline fun <reified VM : ViewModel> Fragment.viewModels(
   noinline ownerProducer: () -> ViewModelStoreOwner = { this },