+++ docs/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change log for Dispatch
 
-## Version 1.0.0-beta08 (unreleased)
+## Version 1.0.0-beta09
+
+* [CoroutineTestExtension] now creates a new [scope][CoroutineTestExtension.scope] instance for each test.
+* Update Kotlin to [1.5.0](https://github.com/JetBrains/kotlin/releases/tag/v1.5.0)
+* Update coroutines to [1.4.3](https://github.com/Kotlin/kotlinx.coroutines/releases/tag/1.4.3)
+* Update Android Gradle Plugin to [4.2.0](https://developer.android.com/studio/releases/gradle-plugin#4-2-0)
+* Update Androidx Lifecycle to [2.3.1](https://developer.android.com/jetpack/androidx/releases/lifecycle#version_231_2)
+
+## Version 1.0.0-beta08
 
 * Remove FlowCancellationException ([#196](https://github.com/RBusarow/Dispatch/issues/196))
 * Update Kotlin to [1.4.21](https://github.com/JetBrains/kotlin/blob/master/ChangeLog.md#1421)
@@ -107,32 +115,45 @@
 * Lower JDK version to 1.6
 
 <!--- MODULE dispatch-core-->
-<!--- INDEX  -->
+<!--- INDEX -->
+
 [DefaultDispatcherProvider]: https://rbusarow.github.io/Dispatch/api/dispatch-core/dispatch.core/-default-dispatcher-provider/index.html
+
 <!--- MODULE dispatch-test-->
 <!--- INDEX  -->
+
 [testProvided]: https://rbusarow.github.io/Dispatch/api/dispatch-test/dispatch.test/test-provided.html
 [TestProvidedCoroutineScope]: https://rbusarow.github.io/Dispatch/api/dispatch-test/dispatch.test/-test-provided-coroutine-scope/index.html
+
 <!--- MODULE dispatch-test-junit4-->
 <!--- INDEX  -->
 <!--- MODULE dispatch-test-junit5-->
 <!--- INDEX  -->
+
 [CoroutineTestExtension]: https://rbusarow.github.io/Dispatch/api/dispatch-test-junit5/dispatch.test/-coroutine-test-extension/index.html
+[CoroutineTestExtension.scope]: https://rbusarow.github.io/Dispatch/api/dispatch-test-junit5/dispatch.test/-coroutine-test-extension/index.html#dispatch.test/CoroutineTestExtension/scope/#/PointingToDeclaration/
+
 <!--- MODULE dispatch-android-espresso-->
 <!--- INDEX  -->
 <!--- MODULE dispatch-android-lifecycle-->
 <!--- INDEX  -->
+
 [ViewLifecycleCoroutineScope]: https://rbusarow.github.io/Dispatch/api/dispatch-android-lifecycle/dispatch.android.lifecycle/-view-lifecycle-coroutine-scope/index.html
 [DispatchLifecycleScope]: https://rbusarow.github.io/Dispatch/api/dispatch-android-lifecycle/dispatch.android.lifecycle/-dispatch-lifecycle-scope/index.html
 [LifecycleCoroutineScope]: https://rbusarow.github.io/Dispatch/api/dispatch-android-lifecycle/dispatch.android.lifecycle/index.html#dispatch.android.lifecycle/LifecycleCoroutineScope//PointingToDeclaration/
+
 <!--- MODULE dispatch-android-lifecycle-extensions-->
 <!--- INDEX  -->
+
 [lifecycleScope]: https://rbusarow.github.io/Dispatch/api/dispatch-android-lifecycle-extensions/dispatch.android.lifecycle/index.html#dispatch.android.lifecycle/lifecycleScope/androidx.lifecycle.LifecycleOwner#/PointingToDeclaration/
+
 <!--- MODULE dispatch-android-viewmodel-->
 <!--- INDEX  -->
+
 [CoroutineViewModel]: https://rbusarow.github.io/Dispatch/api/dispatch-android-viewmodel/dispatch.android.viewmodel/index.html#dispatch.android.viewmodel/CoroutineViewModel//PointingToDeclaration/
 [DispatchViewModel]: https://rbusarow.github.io/Dispatch/api/dispatch-android-viewmodel/dispatch.android.viewmodel/-dispatch-view-model/index.html
 [viewModelScope]: https://rbusarow.github.io/Dispatch/api/dispatch-android-viewmodel/dispatch.android.viewmodel/-dispatch-view-model/index.html#dispatch.android.viewmodel/DispatchViewModel/viewModelScope/#/PointingToDeclaration/
+
 <!--- END -->
 
 [Fragment]: https://developer.android.com/reference/androidx/fragment/app/Fragment